pyproject.toml
[tool.poetry]
name = "boilerplate-auth-payment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
Markdown = "^3.5"
django-filter = "^23.3"
psycopg = "^3.1.12"
python-decouple = "^3.8"
pillow = "^10.0.1"
stripe = "^6.7.0"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.292"
black = "^23.9.1"
pycodestyle = "^2.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
